requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-environ>=0.11.0
django-allauth>=0.57.0
django-ckeditor>=6.7.0
django-storages>=1.14.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
whitenoise>=6.6.0
gunicorn>=21.2.0
python-decouple>=3.8
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
django-taggit>=5.0.0
